fastboot: Add support for "current-slot" variable
It returns the current a.k.a. active a/b slot the OS will boot from.
To avoid code duplication re-use "GptNextKernelEntry" from vboot to find
the active kernel partition. The same function is used by depthcharge
during boot. Note that this logic only works when android is installed
on internal storage, if the device is flashed with ChromeOS this command
will return an error.

BRANCH=none
BUG=b:397322409
TEST=fastboot -s tcp:${dut_ip}:5554 getvar current-slot && \
     fastboot -s tcp:${dut_ip}:5554 set_active b && \
     fastboot -s tcp:${dut_ip}:5554 getvar current-slot && \
     fastboot -s tcp:${dut_ip}:5554 set_active a && \
     fastboot -s tcp:${dut_ip}:5554 getvar current-slot

Cq-Depend: chromium:6297974
Signed-off-by: Kornel Dulęba <korneld@google.com>
Change-Id: Ib645f2693fbe7b454ad681e23d643ed431dca239
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/6300494
Reviewed-by: Tomasz Michalec <tmichalec@google.com>