vboot: Fix NULL pointer dereference in vboot_boot_kernel
The commit addresses a potential NULL pointer dereference in the
vboot_boot_kernel() introduced by the port of commit
a46194fd7c24 ("vboot: Enable booting GKI image") to the
firmware-android-15949.B branch. If the image is not an Android GKI, the
bi.kparams is not initialized and therefore, bi.kparams should not be
used to evaluate if the image is a multiboot image.

This problem was causing hangs in Depthcharge during the warm reset
cycle test on a Fatcat device, with a frequency of 1 out of 5
attempts. The hang signature is provided below.

   multiboot_boot: Multiboot header not found at 0x100000

BRANCH=firmware-android-15949.B
TEST=The warm reset cycle test is able to complete 50 cycles on a Fatcat
     device.

Change-Id: I01d0dcf4d36755fd9188bccbd06e5119587e0e38
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/6715623
Reviewed-by: Karthikeyan Ramasubramanian <kramasub@google.com>
Reviewed-by: Subrata Banik <subratabanik@chromium.org>
Reviewed-by: YH Lin <yueherngl@chromium.org>
Reviewed-by: Bora Guvendik <bora.guvendik@intel.corp-partner.google.com>
Reviewed-by: Zhixing Ma <zhixing.ma@intel.com>
Reviewed-by: Wonkyu Kim <wonkyu.kim@intel.com>